refactor(transport): remove libssh2/SSHTransport, HerdrKit is iOS-buildable (#26)
CitadelTransport (pure-Swift nio-ssh + Citadel) proved out against the
live contract, so the libssh2 path it replaces is deleted: SSHTransport,
CSSH, DescriptorAudit, LiveChannel, SSHRuntime, the fd-ownership thread
pool. This unblocks the iOS build — the libssh2 XCFramework could not
link there, the whole reason Route B exists.

Preserved, not lost:
- SSHCredentials moved to Transport.swift (transport-agnostic).
- The transport-agnostic live contract (agent.list/read/ping parity)
  moved to CitadelTransportContractTests, re-pointed SSH -> Citadel vs Unix.
- The pinning-policy concurrency + scoping tests moved to
  HostKeyPinningTests.
Deleted with the transport: its libssh2-mechanic tests (DescriptorAudit
double-close, LiveChannel adoption, resolver, setup-timeout) — they test
code that no longer exists.

Package.swift drops the CSSH systemLibrary; the symbolgraph and README
layout tests drop their libssh2 references. HerdrKit builds with no
system dependency.

Full suite: 171 tests, 0 failures, -Xswiftc -warnings-as-errors.
