--- Sources/CSSH/module.modulemap
@@ -1,5 +1,0 @@
-module CSSH [system] {
-    header "shim.h"
-    link "ssh2"
-    export *
-}
