--- Sources/HerdrKit/SSHTransport.swift
@@ -1,1161 +1,0 @@
-import CSSH
-import Foundation
-
-#if canImport(Glibc)
-import Glibc
-#elseif canImport(Darwin)
-import Darwin
-#endif
-
-/// How to reach and authenticate to a herdr host.
-public struct SSHCredentials: Sendable {
-    public var host: String
-    public var port: UInt16
-    public var username: String
-    /// PEM private key bytes. Held in memory and handed to
-    /// `libssh2_userauth_publickey_frommemory` — never written to disk and never
-    /// referenced by path, so the key material's lifetime is the caller's to
-    /// control rather than the filesystem's.
-    public var privateKeyPEM: String
-    /// Optional public key bytes. libssh2 derives one when absent.
-    public var publicKeyPEM: String?
-    public var passphrase: String?
-    /// Remote unix socket to forward to, e.g. `~/.config/herdr/herdr.sock`
-    /// resolved to an absolute path on the host.
-    public var remoteSocketPath: String
-
-    public init(
-        host: String,
-        port: UInt16 = 22,
-        username: String,
-        privateKeyPEM: String,
-        publicKeyPEM: String? = nil,
-        passphrase: String? = nil,
-        remoteSocketPath: String
-    ) {
-        self.host = host
-        self.port = port
-        self.username = username
-        self.privateKeyPEM = privateKeyPEM
-        self.publicKeyPEM = publicKeyPEM
-        self.passphrase = passphrase
-        self.remoteSocketPath = remoteSocketPath
-    }
-}
-
-// HostKeyDecision, HostKeyPolicy, PinningHostKeyPolicy and its PinStore moved to
-// HostKeyPinning.swift — the pinning policy is a trust decision shared by every
-// transport, not a libssh2 mechanic, so it must outlive this file.
-
-public enum SSHError: Error, CustomStringConvertible {
-    case resolveFailed(host: String)
-    case connectFailed(host: String, port: UInt16, errno: Int32)
-    /// The setup budget ran out in a named phase. Distinct from a connection
-    /// refusal: nothing failed, the peer simply never answered in time, and the
-    /// caller may want to retry where it would not retry a refusal.
-    case setupTimedOut(phase: String)
-    case sessionInitFailed
-    case handshakeFailed(code: Int32)
-    case hostKeyUnavailable
-    case hostKeyRejected(host: String, reason: String)
-    case authenticationFailed(code: Int32)
-    case channelOpenFailed(socket: String, code: Int32)
-    case writeFailed(code: Int32)
-    case runtimeInitFailed(code: Int32)
-    case readFailed(code: Int32)
-
-    public var description: String {
-        switch self {
-        case .resolveFailed(let h): return "could not resolve \(h)"
-        case .connectFailed(let h, let p, let e): return "connect \(h):\(p) failed (errno \(e))"
-        case .setupTimedOut(let phase): return "SSH setup timed out during \(phase)"
-        case .sessionInitFailed: return "libssh2_session_init failed"
-        case .handshakeFailed(let c): return "SSH handshake failed (\(c))"
-        case .hostKeyUnavailable: return "server presented no host key"
-        case .hostKeyRejected(let h, let r): return "host key rejected for \(h): \(r)"
-        case .authenticationFailed(let c): return "public-key authentication failed (\(c))"
-        case .channelOpenFailed(let s, let c):
-            return "direct-streamlocal to \(s) failed (\(c))"
-        case .writeFailed(let c): return "channel write failed (\(c))"
-        case .runtimeInitFailed(let c): return "libssh2_init failed (\(c))"
-        case .readFailed(let c): return "channel read failed (\(c))"
-        }
-    }
-}
-
-/// Carries herdr's JSON API over an SSH tunnel.
-///
-/// Uses `direct-streamlocal@openssh.com`, not `direct-tcpip`: only the former
-/// can reach a *remote unix socket*, which is where herdr listens. That single
-/// capability is why this is built on libssh2 — SwiftNIO-SSH and Citadel expose
-/// direct-tcpip only.
-///
-/// Conforms to `HerdrTransport` so the two connection shapes stay distinct, as
-/// they must: herdr's command socket answers one request and closes, while
-/// `events.subscribe` holds its connection open. Each `roundTrip` and each
-/// `stream` therefore opens its own SSH channel.
-public struct SSHTransport: HerdrTransport {
-    /// Identity for descriptor auditing, and the reason it is a reference in a
-    /// value type: the audit needs to attribute a double close to ONE logical
-    /// transport, and copies of this struct are the same logical transport.
-    /// Every channel it opens inherits this token, so a test can ask what THIS
-    /// transport did without reading a process-wide tally that every other
-    /// test also writes to.
-    /// Carries its own audit state, so nothing is keyed by an ADDRESS.
-    ///
-    /// The counters and the injectable closer used to live in
-    /// `[ObjectIdentifier: ...]` maps, which are keyed by a pointer value that
-    /// the allocator reuses: review released a token WITHOUT calling forget(),
-    /// obtained a replacement at the same address, and watched the retired
-    /// token's closer run for it. Manual cleanup is not a lifetime; holding the
-    /// state on the object IS one, and it deletes the whole class of hazard
-    /// rather than adding a rule people must remember.
-    final class AuditToken: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _closer: @Sendable (Int32) -> Int32 = { DescriptorAudit.realClose($0) }
-        private var _doubleCloses = 0
-        private var _rejections = 0
-
-        var closer: @Sendable (Int32) -> Int32 {
-            get { lock.withLock { _closer } }
-            set { lock.withLock { _closer = newValue } }
-        }
-        var doubleCloses: Int { lock.withLock { _doubleCloses } }
-        var adoptionRejections: Int { lock.withLock { _rejections } }
-        fileprivate func recordDoubleClose() { lock.withLock { _doubleCloses += 1 } }
-        fileprivate func recordRejection() { lock.withLock { _rejections += 1 } }
-    }
-    let auditToken = AuditToken()
-
-    /// Internal seam: fires once a descriptor exists and BEFORE it is offered to
-    /// the LiveChannel for adoption.
-    ///
-    /// It exists to reach one branch and nothing else. The adopt-rejection
-    /// cleanup runs only when a channel is already closed at that instant, and
-    /// no ordinary cancellation lands there: interrupting BEFORE connect is
-    /// caught earlier, at the resolution claim, and interrupting later lands in
-    /// the catch cleanup instead. Without a hook inside the window the branch is
-    /// unreachable from a test, and a deliberate double close in it survived the
-    /// whole SSH suite.
-    var afterDescriptorPublished: (@Sendable (LiveChannel?) -> Void)?
-
-    /// Shared, bounded pool for blocking libssh2 work. Bounded because a thread
-    /// per request grows without limit under concurrency; shared because each
-    /// request needs its own connection but not its own operating-system thread.
-    /// Stack size is left at the platform default rather than a number picked
-    /// without measuring.
-    static let blockingQueue: OperationQueue = {
-        let q = OperationQueue()
-        q.name = "herdrkit.ssh.blocking"
-        q.maxConcurrentOperationCount = 4
-        return q
-    }()
-
-    public let credentials: SSHCredentials
-    private let hostKeyPolicy: HostKeyPolicy
-    /// Total budget for establishing a session: resolution, connect, handshake
-    /// and authentication share it, and each phase gets what the previous ones
-    /// left.
-    ///
-    /// The default of 20s is **not** a measured value and is not offered as one.
-    /// It is a liveness ceiling — the point past which a phone user is owed an
-    /// error rather than a spinner — chosen to sit far above anything observed:
-    /// `docs/transport-measurements.md` puts handshake plus authentication at
-    /// ~39ms on loopback, with a worst recorded outlier of 2594ms. Tuning for
-    /// latency belongs to the pool work, where checkout cost is measured;
-    /// nothing here should be read as a tuned number.
-    public let setupTimeout: TimeInterval
-
-    /// Test seam: runs after `poll` reports the connect finished and before
-    /// `SO_ERROR` is classified.
-    ///
-    /// That interval is a few instructions wide and is where a cancellation gets
-    /// misread as a peer reset. I claimed it could not be reached deterministically
-    /// and left it uncovered; the reviewer supplied the missing half — a
-    /// bound-but-not-listening loopback port yields EINPROGRESS, then poll
-    /// readiness, then a genuine `SO_ERROR = ECONNREFUSED` (reproduced 5/5). With
-    /// a real post-connect failure available, a hook is all that was missing.
-    ///
-    /// `nil` on every production path.
-    var afterConnectReadyHook: (@Sendable () -> Void)?
-
-    public init(
-        credentials: SSHCredentials,
-        hostKeyPolicy: HostKeyPolicy = PinningHostKeyPolicy(),
-        setupTimeout: TimeInterval = 20
-    ) {
-        self.credentials = credentials
-        self.hostKeyPolicy = hostKeyPolicy
-        self.setupTimeout = SSHTransport.representableSetupTimeout(setupTimeout)
-    }
-
-    /// Below this, millisecond truncation would turn a bound into no bound.
-    public static let minimumSetupTimeout: TimeInterval = 0.001
-
-    /// Above this, the request is a bound in name only.
-    public static let maximumSetupTimeout: TimeInterval = 600
-
-    /// Maps any `TimeInterval` onto a bound libssh2 can actually be given.
-    ///
-    /// Three ways a caller's number stops being a bound, all of which reached
-    /// libssh2 before:
-    ///
-    /// - **Under 1ms** truncates to 0 milliseconds, and libssh2 documents 0 as
-    ///   *no timeout* — so the smallest possible request produced the largest
-    ///   possible behaviour.
-    /// - **NaN** slips through `max(_:_:)` untouched, because every comparison
-    ///   against NaN is false and `max` returns its first argument. `Int(nan *
-    ///   1000)` then traps, so an unusable value became a crash rather than a
-    ///   clamp.
-    /// - **Infinity** has no `Int` representation and traps on conversion.
-    ///
-    /// Non-finite values clamp to the ceiling rather than disabling the bound:
-    /// "wait forever" is exactly the behaviour this transport refuses to expose,
-    /// since an unbounded setup occupies a worker no cancellation can reach.
-    static func representableSetupTimeout(_ requested: TimeInterval) -> TimeInterval {
-        guard requested.isFinite else { return maximumSetupTimeout }
-        return min(max(requested, minimumSetupTimeout), maximumSetupTimeout)
-    }
-
-    /// Remaining budget in milliseconds, floored at 1.
-    ///
-    /// Floored because libssh2 reads 0 as *no timeout*: an exhausted budget
-    /// passed through verbatim would remove the bound at precisely the moment
-    /// it is most needed.
-    static func milliseconds(remainingUntil deadline: Date) -> Int {
-        max(Int(deadline.timeIntervalSinceNow * 1000), 1)
-    }
-
-    /// Bound on teardown **in total**, across every blocking call it makes.
-    ///
-    /// The session timeout is cleared to zero after authentication so an idle
-    /// subscription is never truncated — but zero means *no timeout* for every
-    /// blocking call, including `libssh2_channel_free`, which may wait for the
-    /// peer's close message. Teardown therefore takes its own finite bound: a
-    /// silent peer must not strand a worker after `roundTrip` has already
-    /// resumed its caller.
-    ///
-    /// Stated as a total because the first version was not one. libssh2's
-    /// timeout applies per blocking call, so setting it once gave
-    /// `channel_free` and the session disconnect two seconds *each* — a 2s
-    /// constant that permitted 4s, which a passing test with a three-second
-    /// margin cheerfully hid. Teardown now carries one deadline and re-arms
-    /// libssh2 with what remains before each call.
-    public static let teardownTimeout: TimeInterval = 2
-
-    // MARK: - Session
-
-    /// An authenticated session plus the socket beneath it. Both are torn down
-    /// together; the caller owns the lifetime.
-    final class Session {
-        let sock: Int32
-        let handle: OpaquePointer
-        /// Who a double close here is attributed to. See DescriptorAudit.
-        var auditOwner: SSHTransport.AuditToken?
-
-        init(sock: Int32, handle: OpaquePointer) {
-            self.sock = sock
-            self.handle = handle
-        }
-
-        /// Everything except the descriptor.
-        ///
-        /// Split out so `LiveChannel` can keep the socket shuttable-down through
-        /// the blocking part of teardown and still stop publishing the number
-        /// *before* it is closed. Closing it while it is still published leaves a
-        /// window where a concurrent `interrupt()` shuts down whatever the
-        /// process has since opened on that number — the double-close hazard
-        /// running the other way.
-        func disconnectAndFree() {
-            libssh2_session_disconnect_ex(handle, SSH_DISCONNECT_BY_APPLICATION, "bye", "")
-            libssh2_session_free(handle)
-        }
-
-        func closeSocket() {
-            // THROUGH THE AUDIT. This closed the descriptor directly, so the
-            // one test asserting "LiveChannel never closes a descriptor
-            // openSession already closed" drove a path the audit could not see:
-            // on any route where a session exists — host-key rejection among
-            // them — teardown comes here, not to the audited branch. A
-            // deliberate double close in that branch SURVIVED the test.
-            // The assertion was vacuous, and its process-wide counter then made
-            // it flaky on other tests' noise: a meter measuring nothing,
-            // wobbling.
-            DescriptorAudit.close(sock, owner: auditOwner)
-        }
-
-        func close() {
-            disconnectAndFree()
-            closeSocket()
-        }
-    }
-
-    /// One resolved address, copied out of `addrinfo` as a value.
-    ///
-    /// Copying rather than passing the list through removes the ownership
-    /// question entirely: `freeaddrinfo` runs on the resolver thread the moment
-    /// the copy is made, so there is no allocation whose owner depends on who
-    /// gave up first. The previous version got that handoff right, but only by
-    /// having both sides check a flag under the same lock — correct, and one
-    /// edit away from not being.
-    struct ResolvedAddress {
-        var family: Int32
-        var socketType: Int32
-        var protocolNumber: Int32
-        var storage: sockaddr_storage
-        var length: socklen_t
-    }
-
-    /// Name resolution running off the caller's thread, so a stalled resolver
-    /// strands a throwaway thread instead of a pool worker.
-    ///
-    /// `getaddrinfo` has no timeout and no cancellation: once called, it returns
-    /// when the resolver decides to. On a queue of four workers that is
-    /// unbounded occupancy, and no interrupt can reach it — there is no
-    /// descriptor to shut down yet, which is what makes this different from
-    /// every other blocking call here.
-    ///
-    /// Waiters use `NSCondition` rather than a semaphore because a resolution is
-    /// shared: a semaphore signalled once wakes one waiter, and every other
-    /// caller attached to the same lookup would sit until its deadline for an
-    /// answer that had already arrived.
-    final class Resolution: @unchecked Sendable {
-        enum Outcome {
-            case resolved([ResolvedAddress])
-            case failed(Int32)
-            case gaveUp
-        }
-
-        private let condition = NSCondition()
-        private var addresses: [ResolvedAddress] = []
-        private var code: Int32 = 0
-        private var finished = false
-
-        /// How many resolver threads this process has started.
-        ///
-        /// Counting *constructions* rather than the objects callers receive is
-        /// the whole point: a check-create-check registry returns one shared
-        /// object while having started and discarded fifteen threads, so a test
-        /// that counts what callers got cannot see the bug. One did not, and the
-        /// mutation restoring that shape survived it.
-        private static let auditLock = NSLock()
-        private static var constructions = 0
-
-        static var constructionCount: Int {
-            auditLock.lock(); defer { auditLock.unlock() }
-            return constructions
-        }
-
-        /// Test seam: runs on the resolver thread before `getaddrinfo`, so a
-        /// test can hold a resolution in flight and force callers to overlap.
-        ///
-        /// Read and written under `auditLock`, and **captured at construction**
-        /// rather than read from the resolver thread. The first version was an
-        /// unsynchronised `static var`: resolver threads read it while the test's
-        /// `defer` cleared it, which ThreadSanitizer caught twice even though ten
-        /// ordinary runs passed. A seam whose own guard is undefined behaviour
-        /// cannot pin anything.
-        nonisolated(unsafe) private static var _startGate: (@Sendable () -> Void)?
-
-        static var startGate: (@Sendable () -> Void)? {
-            get { auditLock.lock(); defer { auditLock.unlock() }; return _startGate }
-            set { auditLock.lock(); _startGate = newValue; auditLock.unlock() }
-        }
-
-        init(host: String, port: UInt16, onFinish: @escaping @Sendable (Resolution) -> Void) {
-            Resolution.auditLock.lock()
-            Resolution.constructions += 1
-            // Captured here, so the resolver thread invokes an immutable local
-            // and never touches the shared property — and the lock is released
-            // before the gate can block on it.
-            let gate = Resolution._startGate
-            Resolution.auditLock.unlock()
-            let thread = Thread { [self] in
-                gate?()
-                var hints = addrinfo()
-                hints.ai_family = AF_UNSPEC
-                hints.ai_socktype = sockStream
-                var info: UnsafeMutablePointer<addrinfo>?
-                let rc = getaddrinfo(host, String(port), &hints, &info)
-
-                var copied: [ResolvedAddress] = []
-                var candidate = info
-                while let entry = candidate {
-                    if let address = entry.pointee.ai_addr {
-                        var storage = sockaddr_storage()
-                        let length = min(
-                            Int(entry.pointee.ai_addrlen), MemoryLayout<sockaddr_storage>.size
-                        )
-                        _ = withUnsafeMutableBytes(of: &storage) { destination in
-                            UnsafeRawBufferPointer(start: address, count: length)
-                                .copyBytes(to: destination)
-                        }
-                        copied.append(ResolvedAddress(
-                            family: entry.pointee.ai_family,
-                            socketType: entry.pointee.ai_socktype,
-                            protocolNumber: entry.pointee.ai_protocol,
-                            storage: storage,
-                            length: socklen_t(length)
-                        ))
-                    }
-                    candidate = entry.pointee.ai_next
-                }
-                if let info { freeaddrinfo(info) }
-
-                condition.lock()
-                addresses = copied
-                code = rc
-                finished = true
-                condition.broadcast()
-                condition.unlock()
-                onFinish(self)
-            }
-            thread.name = "herdrkit.ssh.resolve"
-            thread.start()
-        }
-
-        var isFinished: Bool {
-            condition.lock(); defer { condition.unlock() }
-            return finished
-        }
-
-        /// Waits in slices so an interrupt is noticed without sitting out the
-        /// whole deadline. Giving up costs nothing — there is no allocation to
-        /// hand back, and the resolver thread finishes on its own.
-        func claim(by deadline: Date, isInterrupted: () -> Bool) -> Outcome {
-            condition.lock(); defer { condition.unlock() }
-            while !finished {
-                if isInterrupted() { return .gaveUp }
-                let remaining = deadline.timeIntervalSinceNow
-                guard remaining > 0 else { return .gaveUp }
-                _ = condition.wait(until: Date().addingTimeInterval(min(0.02, remaining)))
-            }
-            if !addresses.isEmpty { return .resolved(addresses) }
-            return .failed(code)
-        }
-    }
-
-    /// One in-flight resolution per endpoint, shared by every caller waiting on it.
-    ///
-    /// Without this, abandoning a stalled resolver *relocates* exhaustion rather
-    /// than bounding it: each timed-out request leaves its own detached thread,
-    /// so a wedged DNS server plus a retrying client grows threads without
-    /// limit. Sharing caps it at one resolver thread per distinct endpoint —
-    /// and this transport has exactly one endpoint, so a wedged resolver costs
-    /// one thread no matter how many requests pile up behind it.
-    ///
-    /// Entries are removed once resolution completes, so the next request
-    /// resolves afresh. This is a concurrency bound, not a DNS cache; caching
-    /// answers would need a TTL and would hide a changed address.
-    enum ResolutionRegistry {
-        private static let lock = NSLock()
-        private static var inFlight: [String: Resolution] = [:]
-
-        /// Constructed **while holding the lock**, deliberately.
-        ///
-        /// A check-then-create outside the lock still lets sixteen simultaneous
-        /// first-callers each start a resolver thread and then discard fifteen —
-        /// the registry would report one entry while sixteen threads ran, which
-        /// is the exhaustion this exists to prevent, invisible.
-        ///
-        /// Safe against the completion callback because the resolver thread
-        /// releases the condition lock *before* calling `onFinish`, so it never
-        /// holds a lock this path is waiting on. If that ever changes, this
-        /// deadlocks — the ordering is load-bearing, not incidental.
-        static func resolution(host: String, port: UInt16) -> Resolution {
-            let key = "\(host):\(port)"
-            lock.lock(); defer { lock.unlock() }
-            if let existing = inFlight[key], !existing.isFinished { return existing }
-            let fresh = Resolution(host: host, port: port) { completed in
-                lock.lock()
-                if inFlight[key] === completed { inFlight[key] = nil }
-                lock.unlock()
-            }
-            inFlight[key] = fresh
-            return fresh
-        }
-
-        /// In-flight resolutions, for tests asserting the bound holds.
-        static var inFlightCount: Int {
-            lock.lock(); defer { lock.unlock() }
-            return inFlight.count
-        }
-    }
-
-    /// Resolves and connects within `deadline`, publishing every candidate
-    /// descriptor before it can block on that descriptor.
-    ///
-    /// The previous version could not be bounded or cancelled at all: it
-    /// resolved and then called blocking `connect`, both before the socket was
-    /// handed to `LiveChannel`. A SYN to a host that accepts nothing holds the
-    /// worker for the kernel's retry schedule — minutes — and four of those
-    /// exhaust the queue while `setupTimeout` and task cancellation both look
-    /// on. Non-blocking connect plus `poll` turns that into the deadline the
-    /// caller actually asked for.
-    private func tcpConnect(by deadline: Date, publishingTo live: LiveChannel?) throws -> Int32 {
-        let resolution = ResolutionRegistry.resolution(host: credentials.host, port: credentials.port)
-        let addresses: [ResolvedAddress]
-        switch resolution.claim(by: deadline, isInterrupted: { live?.isInterrupted ?? false }) {
-        case .resolved(let resolved):
-            addresses = resolved
-        case .failed:
-            throw SSHError.resolveFailed(host: credentials.host)
-        case .gaveUp:
-            if live?.isInterrupted == true { throw CancellationError() }
-            throw SSHError.setupTimedOut(phase: "resolve")
-        }
-
-        var lastErrno: Int32 = 0
-        for address in addresses {
-            let fd = socket(address.family, address.socketType, address.protocolNumber)
-            guard fd >= 0 else { lastErrno = errno; continue }
-            let generation = DescriptorAudit.opened(fd)
-            afterDescriptorPublished?(live)
-            // Published before the first call that can block on it, so an
-            // interrupt arriving mid-connect has a real descriptor to shut down.
-            guard live?.adopt(rawSocket: fd) ?? true else {
-                DescriptorAudit.closeRejectedAdoption(fd, generation: generation, owner: auditToken)
-                throw CancellationError()
-            }
-            do {
-                try connectCandidate(fd, address, by: deadline, live: live)
-                // SSH carries small request/response exchanges; Nagle holds a
-                // small write waiting for an ACK the peer has delayed, which
-                // shows up as a fixed ~40ms floor rather than as work.
-                var on: Int32 = 1
-                setsockopt(fd, Int32(IPPROTO_TCP), TCP_NODELAY, &on, socklen_t(MemoryLayout<Int32>.size))
-                return fd
-            } catch {
-                // This candidate is ours to clean up, so relinquish the
-                // published number before closing it — otherwise `close()` closes
-                // it a second time, by which point it may name another socket.
-                live?.release()
-                DescriptorAudit.close(fd, owner: auditToken)
-                guard case SSHError.connectFailed(_, _, let failure) = error else {
-                    // A deadline or an interrupt ends the attempt outright: both
-                    // are global, and trying the next address cannot help.
-                    throw error
-                }
-                lastErrno = failure
-            }
-        }
-        // Same reasoning one level up: the last candidate's failure may itself
-        // have been our own shutdown.
-        if live?.isInterrupted == true { throw CancellationError() }
-        throw SSHError.connectFailed(host: credentials.host, port: credentials.port, errno: lastErrno)
-    }
-
-    private func connectCandidate(
-        _ fd: Int32, _ address: ResolvedAddress, by deadline: Date, live: LiveChannel?
-    ) throws {
-        func failed(_ code: Int32) -> SSHError {
-            .connectFailed(host: credentials.host, port: credentials.port, errno: code)
-        }
-
-        let flags = fcntl(fd, F_GETFL, 0)
-        guard flags >= 0, fcntl(fd, F_SETFL, flags | O_NONBLOCK) == 0 else { throw failed(errno) }
-
-        var storage = address.storage
-        let attempt = withUnsafePointer(to: &storage) { pointer in
-            pointer.withMemoryRebound(to: sockaddr.self, capacity: 1) {
-                connect(fd, $0, address.length)
-            }
-        }
-        if attempt != 0 {
-            let pending = errno
-            guard pending == EINPROGRESS || pending == EINTR else { throw failed(pending) }
-            try waitWritable(fd, by: deadline, live: live)
-            afterConnectReadyHook?()
-            // poll() reporting the socket writable says the attempt finished,
-            // not that it succeeded — a refusal is also a completion. SO_ERROR
-            // carries which one it was.
-            var pendingError: Int32 = 0
-            var size = socklen_t(MemoryLayout<Int32>.size)
-            guard getsockopt(fd, SOL_SOCKET, SO_ERROR, &pendingError, &size) == 0 else {
-                throw failed(errno)
-            }
-            // Deliberately does NOT recheck interruption here, though a
-            // cancellation landing in this window does make SO_ERROR report a
-            // failure. The recheck at the end of `tcpConnect` already covers it,
-            // and mutation showed neither guard was individually pinned: either
-            // one alone passed the test, and only removing both failed it. That
-            // is the same trap `verifyHostKey` above records — two guards for one
-            // property means the test pins neither, and a mutation of the
-            // redundant one reports a coverage gap that does not exist.
-            guard pendingError == 0 else { throw failed(pendingError) }
-        }
-
-        // libssh2 is used in blocking mode, so hand it back a blocking
-        // descriptor: non-blocking was only ever a device for bounding connect.
-        guard fcntl(fd, F_SETFL, flags) == 0 else { throw failed(errno) }
-    }
-
-    /// Waits for a connect to finish, in slices, until the deadline.
-    ///
-    /// The interrupt check is **not** what unblocks a cancelled connect —
-    /// measured, not assumed: `shutdown()` on a still-connecting socket wakes
-    /// `poll` on Linux within a millisecond (`revents` = POLLOUT|POLLERR|POLLHUP,
-    /// `SO_ERROR` = ECONNRESET). What the check provides is the *right error*.
-    /// Without it, a cancelled connect surfaces as `connectFailed(ECONNRESET)`,
-    /// indistinguishable from a peer that reset us — and a caller that retries
-    /// connection failures would then retry a request the user cancelled.
-    ///
-    /// Sliced so the check is reached promptly even where that wakeup does not
-    /// hold, rather than depending on one platform's behaviour for correctness.
-    private func waitWritable(_ fd: Int32, by deadline: Date, live: LiveChannel?) throws {
-        while true {
-            if live?.isInterrupted == true { throw CancellationError() }
-            let remaining = deadline.timeIntervalSinceNow
-            guard remaining > 0 else { throw SSHError.setupTimedOut(phase: "connect") }
-            var descriptor = pollfd(fd: fd, events: Int16(POLLOUT), revents: 0)
-            let slice = Int32(max(min(remaining, 0.05) * 1000, 1))
-            let ready = poll(&descriptor, 1, slice)
-            if ready < 0 {
-                if errno == EINTR { continue }
-                throw SSHError.connectFailed(
-                    host: credentials.host, port: credentials.port, errno: errno
-                )
-            }
-            if ready > 0 {
-                // Readiness is not necessarily a finished connect. Shutting the
-                // socket down to cancel makes poll ready too, with SO_ERROR set
-                // to ECONNRESET — so returning here unasked turns a cancellation
-                // into a peer reset. Checking only at the top of the loop missed
-                // this entirely, because the wakeup arrives inside the slice.
-                if live?.isInterrupted == true { throw CancellationError() }
-                return
-            }
-        }
-    }
-
-    /// SHA-256 of the presented host key, hex, lowercase.
-    private func fingerprint(_ session: OpaquePointer) throws -> String {
-        guard let raw = libssh2_hostkey_hash(session, LIBSSH2_HOSTKEY_HASH_SHA256) else {
-            throw SSHError.hostKeyUnavailable
-        }
-        return (0..<32).map { String(format: "%02x", UInt8(bitPattern: raw[$0])) }.joined()
-    }
-
-    private func verifyHostKey(_ session: OpaquePointer) throws {
-        let presented = try fingerprint(session)
-        // One decision, one rejection path. An earlier version rejected in two
-        // places, which made the first branch redundant — and a mutation of it
-        // left the test green, because the second branch was doing the work.
-        // Two guards for one property means neither is pinned by its test.
-        // One call: the policy compares and pins atomically. A split
-        // lookup-then-decide interface leaves a window where two concurrent
-        // first contacts both observe no pin, trust different keys, and
-        // overwrite each other — silently defeating hard-stop-on-change at the
-        // exact moment pinning is meant to establish trust.
-        let decision = hostKeyPolicy.evaluate(
-            host: credentials.host, port: credentials.port, presented: presented
-        )
-        guard decision == .trust else {
-            throw SSHError.hostKeyRejected(
-                host: credentials.host, reason: "host key not trusted for this host:port"
-            )
-        }
-    }
-
-    /// Establishes a session, publishing the raw socket to `live` BEFORE any
-    /// blocking call so cancellation can interrupt establishment itself.
-    ///
-    /// The previous shape completed connect, handshake, host-key check and auth
-    /// and only then handed the socket over — so a cancel arriving during those
-    /// steps had nothing to shut down. A peer that accepts TCP and withholds its
-    /// banner would hold a worker indefinitely, and four such peers exhaust the
-    /// whole queue.
-    func openSession(publishingTo live: LiveChannel? = nil) throws -> Session {
-        try SSHRuntime.ensureStarted()
-        // One budget for the whole of establishment. Each phase gets what the
-        // ones before it left, so a slow resolver cannot buy the handshake a
-        // fresh 20 seconds and the caller's bound means what it says.
-        let deadline = Date().addingTimeInterval(setupTimeout)
-        // tcpConnect publishes each candidate before it can block on it, so by
-        // the time it returns, ownership is already with `live`.
-        let sock = try tcpConnect(by: deadline, publishingTo: live)
-        if let live, live.isInterrupted {
-            live.release()
-            DescriptorAudit.close(sock, owner: auditToken)
-            throw CancellationError()
-        }
-        // SO_RCVTIMEO/SO_SNDTIMEO do NOT bound libssh2 in blocking mode — a
-        // reviewer measured establishment still blocked after 17s against a peer
-        // that withheld its banner, so the 15s bound previously claimed here was
-        // simply false. libssh2's own session timeout does bound it; it is set
-        // for setup and cleared after authentication so it never truncates a
-        // long-lived subscription.
-        guard let handle = libssh2_session_init_ex(nil, nil, nil, nil) else {
-            live?.release()
-            DescriptorAudit.close(sock, owner: auditToken)
-            throw SSHError.sessionInitFailed
-        }
-        libssh2_session_set_blocking(handle, 1)
-
-        func fail(_ error: SSHError) -> SSHError {
-            libssh2_session_free(handle)
-            live?.release()          // we close it here; LiveChannel must not repeat it
-            DescriptorAudit.close(sock, owner: auditToken)
-            return error
-        }
-
-        libssh2_session_set_timeout(handle, SSHTransport.milliseconds(remainingUntil: deadline))
-        let hs = libssh2_session_handshake(handle, sock)
-        guard hs == 0 else { throw fail(.handshakeFailed(code: hs)) }
-
-        do {
-            try verifyHostKey(handle)
-        } catch {
-            libssh2_session_free(handle)
-            live?.release()
-            DescriptorAudit.close(sock, owner: auditToken)
-            throw error
-        }
-
-        // Re-armed with what is left. libssh2's timeout bounds each blocking
-        // call individually, not the session's total, so leaving the handshake's
-        // value in place would hand authentication a second full budget.
-        libssh2_session_set_timeout(handle, SSHTransport.milliseconds(remainingUntil: deadline))
-        let auth = credentials.privateKeyPEM.withCString { priv -> Int32 in
-            let privLen = strlen(priv)
-            let run: (UnsafePointer<CChar>?, Int) -> Int32 = { pub, pubLen in
-                credentials.username.withCString { user in
-                    libssh2_userauth_publickey_frommemory(
-                        handle, user, strlen(user),
-                        pub, pubLen,
-                        priv, privLen,
-                        credentials.passphrase
-                    )
-                }
-            }
-            if let pub = credentials.publicKeyPEM {
-                return pub.withCString { run($0, strlen($0)) }
-            }
-            return run(nil, 0)
-        }
-        guard auth == 0 else { throw fail(.authenticationFailed(code: auth)) }
-
-        // Cleared so an intentionally idle event stream is not truncated. This
-        // DOES reopen unbounded blocking for channel open, write and read — the
-        // reviewer confirmed it against libssh2's own documentation. That is
-        // accepted deliberately: a subscription must be able to sit silent for
-        // minutes, and the interrupt path (LiveChannel shutting the socket down)
-        // is what bounds those operations instead of a timer.
-        libssh2_session_set_timeout(handle, 0)
-        let session = Session(sock: sock, handle: handle)
-        session.auditOwner = auditToken
-        return session
-    }
-
-    /// Opens a direct-streamlocal channel to herdr's socket on the host.
-    func openChannel(_ session: Session) throws -> OpaquePointer {
-        guard let channel = credentials.remoteSocketPath.withCString({ path in
-            libssh2_channel_direct_streamlocal_ex(session.handle, path, "localhost", 0)
-        }) else {
-            throw SSHError.channelOpenFailed(
-                socket: credentials.remoteSocketPath,
-                code: libssh2_session_last_errno(session.handle)
-            )
-        }
-        return channel
-    }
-
-    // MARK: - Channel IO
-
-    private func write(_ channel: OpaquePointer, _ line: String) throws {
-        var bytes = Array(line.utf8)
-        if bytes.last != UInt8(ascii: "\n") { bytes.append(UInt8(ascii: "\n")) }
-        var offset = 0
-        while offset < bytes.count {
-            let n = bytes.withUnsafeBytes { raw -> Int in
-                libssh2_channel_write_ex(
-                    channel, 0,
-                    raw.baseAddress!.advanced(by: offset).assumingMemoryBound(to: CChar.self),
-                    bytes.count - offset
-                )
-            }
-            if n > 0 {
-                offset += n
-            } else if n == Int(LIBSSH2_ERROR_EAGAIN) {
-                continue
-            } else {
-                throw SSHError.writeFailed(code: Int32(n))
-            }
-        }
-    }
-
-    /// Reads to the next newline. Returns nil at channel EOF with nothing buffered.
-    private func readLine(_ channel: OpaquePointer, carry: inout [UInt8]) throws -> String? {
-        while true {
-            if let idx = carry.firstIndex(of: UInt8(ascii: "\n")) {
-                let line = Array(carry[carry.startIndex..<idx])
-                carry.removeSubrange(carry.startIndex...idx)
-                return String(decoding: line, as: UTF8.self)
-            }
-            var buf = [CChar](repeating: 0, count: 16 * 1024)
-            let n = buf.withUnsafeMutableBufferPointer { p -> Int in
-                libssh2_channel_read_ex(channel, 0, p.baseAddress!, p.count)
-            }
-            if n > 0 {
-                carry.append(contentsOf: buf[0..<n].map { UInt8(bitPattern: $0) })
-            } else if n == Int(LIBSSH2_ERROR_EAGAIN) {
-                continue
-            } else if n == 0 {
-                // A zero-byte read is NOT proof of EOF for libssh2 — it also
-                // occurs on a live channel with nothing available. Ask the
-                // channel. Treating zero as EOF ended streams that were merely
-                // idle, which on an event subscription is most of the time.
-                // libssh2_channel_eof: 1 = EOF, 0 = not EOF, NEGATIVE = query
-                // failed. Continuing only on exactly 0 meant every negative —
-                // i.e. every failure — was silently treated as a clean EOF.
-                let eof = libssh2_channel_eof(channel)
-                if eof == 0 { continue }
-                if eof < 0 { throw SSHError.readFailed(code: eof) }
-                if carry.isEmpty { return nil }
-                let rest = String(decoding: carry, as: UTF8.self)
-                carry.removeAll()
-                return rest
-            } else {
-                // A negative code other than EAGAIN is a failure, not a clean
-                // close. Treating it as EOF turned network and protocol errors
-                // into normal stream completion, so a broken connection looked
-                // like a server that had finished talking.
-                throw SSHError.readFailed(code: Int32(n))
-            }
-        }
-    }
-
-    // MARK: - HerdrTransport
-
-    public func roundTrip(_ requestLine: String) async throws -> String {
-        // Blocking work needs a real thread — Task.detached does not isolate it —
-        // but a thread PER REQUEST is unbounded and uncancellable. Concurrent
-        // callers would each allocate an OS thread that keeps connecting,
-        // handshaking or reading long after the awaiting task is gone.
-        //
-        // So: a shared bounded queue, plus a cancellation handle that publishes
-        // the socket so an in-flight blocking read can actually be interrupted.
-        let live = LiveChannel()
-        live.auditOwner = auditToken
-        return try await withTaskCancellationHandler {
-            try await withCheckedThrowingContinuation { continuation in
-                SSHTransport.blockingQueue.addOperation {
-                    do {
-                        // A queued operation may have been cancelled while it
-                        // waited for a worker; do not begin setup at all.
-                        guard !live.isInterrupted else { throw CancellationError() }
-                        let session = try openSession(publishingTo: live)
-                        guard live.adopt(session) else {
-                            session.close()
-                            throw CancellationError()
-                        }
-                        defer { live.close() }
-                        let channel = try openChannel(session)
-                        live.adopt(channel: channel)
-
-                        try write(channel, requestLine)
-                        var carry: [UInt8] = []
-                        guard let line = try readLine(channel, carry: &carry) else {
-                            throw TransportError.closedBeforeResponse
-                        }
-                        continuation.resume(returning: line)
-                    } catch {
-                        continuation.resume(throwing: error)
-                    }
-                }
-            }
-        } onCancel: {
-            // Shuts the socket down so a blocked read returns now, rather than
-            // leaving the worker running until the peer speaks.
-            live.interrupt()
-        }
-    }
-
-    public func stream(_ requestLine: String) -> AsyncThrowingStream<String, Error> {
-        AsyncThrowingStream { continuation in
-            let live = LiveChannel()
-        live.auditOwner = auditToken
-            // Same reasoning as roundTrip: a detached task is not a thread, and
-            // this loop blocks for the life of the subscription.
-            let work = Thread {
-                do {
-                    guard !live.isInterrupted else {
-                        continuation.finish()
-                        return
-                    }
-                    let session = try openSession(publishingTo: live)
-                    guard live.adopt(session) else {
-                        session.close()
-                        continuation.finish()
-                        return
-                    }
-                    let channel = try openChannel(session)
-                    live.adopt(channel: channel)
-                    try write(channel, requestLine)
-
-                    var carry: [UInt8] = []
-                    while !live.isInterrupted, let line = try readLine(channel, carry: &carry) {
-                        continuation.yield(line)
-                    }
-                    live.close()
-                    continuation.finish()
-                } catch {
-                    live.close()
-                    continuation.finish(throwing: error)
-                }
-            }
-            work.name = "herdrkit.ssh.stream"
-            work.start()
-            continuation.onTermination = { _ in
-                // Same discipline as UnixSocketTransport: Task.cancel() cannot
-                // interrupt a blocking read, so the underlying socket is shut
-                // down to force it to return. Without this a cancelled
-                // subscription holds its descriptor and a pool thread until the
-                // server next writes — which on an idle stream may be never
-                // (herdr-ios#1).
-                live.interrupt()
-            }
-        }
-    }
-}
-
-/// Owns a session and channel shared between a blocking reader and a cancelling
-/// caller, so cancellation can interrupt a read that is already in progress.
-final class LiveChannel: @unchecked Sendable {
-    /// Who to attribute a double close to. The channel is created per stream
-    /// and never held by a caller, so attributing to `self` gives a test
-    /// nothing it can ask about; the TRANSPORT outlives every channel it makes
-    /// and is what a test actually holds.
-    var auditOwner: SSHTransport.AuditToken?
-    private let lock = NSLock()
-    private var session: SSHTransport.Session?
-    private var channel: OpaquePointer?
-    private var rawSocket: Int32 = -1
-    /// Set only while `close()` tears down outside the lock, so an interrupt
-    /// arriving during teardown still has something to shut down.
-    private var teardownSocket: Int32 = -1
-    private var closed = false
-
-    /// Signalled as `close()` enters libssh2's teardown.
-    ///
-    /// Internal, and here for one reason: timing `interrupt()` against a sleep
-    /// is a race, and the test that did so passed and failed on alternate runs
-    /// against the same code. A semaphore needs no lock to observe, so it
-    /// reports teardown has begun even in the broken shape where the lock is
-    /// held throughout — which is exactly the shape the test has to distinguish.
-    let teardownBegan = DispatchSemaphore(value: 0)
-
-    /// Relinquish ownership of the published descriptor without closing it,
-    /// for failure paths that close it themselves.
-    ///
-    /// Without this, a failure path closed the fd while LiveChannel still held
-    /// the number, and close() closed it a second time — by which point the
-    /// process may have reused that descriptor for an unrelated socket. A
-    /// host-key rejection probe reproduced exactly that.
-    func release() {
-        lock.lock(); defer { lock.unlock() }
-        rawSocket = -1
-    }
-
-    /// Publish a bare descriptor before a Session exists, so establishment is
-    /// interruptible rather than only the read loop.
-    func adopt(rawSocket fd: Int32) -> Bool {
-        lock.lock(); defer { lock.unlock() }
-        guard !closed else { return false }
-        rawSocket = fd
-        return true
-    }
-
-    /// The Session owns the descriptor from here **whether or not the adoption
-    /// is accepted**.
-    ///
-    /// Clearing only on the accepted path left the rejected one publishing a
-    /// number the caller was about to close — both callers do `session.close()`
-    /// and then throw. A later `interrupt()` would then shut down whatever the
-    /// process had opened on that number since. Same hazard as the double close,
-    /// reached through the cancellation path instead of the failure path.
-    func adopt(_ session: SSHTransport.Session) -> Bool {
-        lock.lock(); defer { lock.unlock() }
-        rawSocket = -1
-        guard !closed else { return false }
-        self.session = session
-        return true
-    }
-
-    func adopt(channel: OpaquePointer) {
-        lock.lock(); defer { lock.unlock() }
-        self.channel = channel
-    }
-
-    /// Unblocks a pending read without freeing anything the reader still holds.
-    var isInterrupted: Bool {
-        lock.lock(); defer { lock.unlock() }
-        return closed
-    }
-
-    func interrupt() {
-        lock.lock(); defer { lock.unlock() }
-        closed = true
-        // Any phase may be in flight: establishment (raw socket published, no
-        // Session yet), the read loop (Session owns it), or teardown (detached,
-        // but still worth unblocking).
-        if let sock = session?.sock { _ = Glibc_shutdownSocket(sock) }
-        else if rawSocket >= 0 { _ = Glibc_shutdownSocket(rawSocket) }
-        else if teardownSocket >= 0 { _ = Glibc_shutdownSocket(teardownSocket) }
-    }
-
-    /// Detaches everything under the lock, then tears down outside it.
-    ///
-    /// Holding the mutex across teardown deadlocked the interrupt path against
-    /// the very stall it exists to break: `libssh2_channel_free` and the session
-    /// disconnect can wait for the peer, the post-auth timeout is zero so they
-    /// wait without bound, and `interrupt()` could not take the lock to shut the
-    /// socket down. A silent peer therefore held a worker after `roundTrip` had
-    /// already resumed its caller.
-    ///
-    /// Two things fix it together, because either alone still fails:
-    /// detaching frees the lock so an interrupt can land, and the finite
-    /// teardown bound covers the case where no interrupt ever comes.
-    func close() {
-        lock.lock()
-        closed = true
-        let channel = self.channel
-        let session = self.session
-        let raw = self.rawSocket
-        self.channel = nil
-        self.session = nil
-        self.rawSocket = -1
-        // Kept shuttable-down while teardown runs unlocked: a concurrent
-        // interrupt must still be able to unblock a teardown stalling on a
-        // peer that has stopped answering.
-        teardownSocket = session?.sock ?? raw
-        lock.unlock()
-
-        teardownBegan.signal()
-        // One deadline for all of teardown, re-armed before each blocking call,
-        // because libssh2's timeout bounds a call rather than a sequence.
-        let deadline = Date().addingTimeInterval(SSHTransport.teardownTimeout)
-        if let session {
-            libssh2_session_set_timeout(
-                session.handle, SSHTransport.milliseconds(remainingUntil: deadline)
-            )
-        }
-        if let channel { libssh2_channel_free(channel) }
-        if let session {
-            libssh2_session_set_timeout(
-                session.handle, SSHTransport.milliseconds(remainingUntil: deadline)
-            )
-            session.disconnectAndFree()
-        }
-        // Stop publishing the number BEFORE anything closes it. Clearing it
-        // afterwards left a window in which `interrupt()` would shut down a
-        // descriptor this method had already closed — and by then the number may
-        // name an unrelated socket, so the shutdown lands on a live connection
-        // somewhere else entirely. Same defect as the double close, running in
-        // the other direction, and no amount of care at the call sites fixes it:
-        // the ordering has to.
-        lock.lock()
-        teardownSocket = -1
-        lock.unlock()
-
-        if let session {
-            session.closeSocket()
-        } else if raw >= 0 {
-            DescriptorAudit.close(raw, owner: auditOwner)
-        }
-    }
-}
-
-/// Counts descriptors closed that were already closed.
-///
-/// A double close is otherwise invisible from a test: `close(2)` returns EBADF
-/// and nothing reads it, and by the time it happens the number may already name
-/// an unrelated socket — which is the actual damage, and it lands somewhere else
-/// entirely. This exists because a double close shipped here once behind a test
-/// whose name said it was guarded, and nothing in the suite could have noticed.
-enum DescriptorAudit {
-    private static let lock = NSLock()
-    private static var doubleCloseCount = 0
-    private static var generations: [Int32: Int] = [:]
-
-    static func realClose(_ fd: Int32) -> Int32 { Glibc_close(fd) }
-
-    /// Closes, and records a double close against the token that owned it.
-    static func close(_ fd: Int32, owner: SSHTransport.AuditToken?) {
-        guard (owner?.closer ?? realClose)(fd) != 0, errno == EBADF else { return }
-        lock.lock(); doubleCloseCount += 1; lock.unlock()
-        owner?.recordDoubleClose()
-    }
-
-    /// Process-wide tally. NOT sound for a test assertion — any concurrent
-    /// teardown lands inside the window. Ask a token instead.
-    static var doubleCloses: Int {
-        lock.lock(); defer { lock.unlock() }
-        return doubleCloseCount
-    }
-
-    /// Records that a descriptor NUMBER has just been opened, returning the
-    /// generation that identifies this particular use of it. Numbers recycle,
-    /// so EBADF alone can never say which opening a close belonged to.
-    static func opened(_ fd: Int32) -> Int {
-        lock.lock(); defer { lock.unlock() }
-        generations[fd, default: 0] += 1
-        return generations[fd] ?? 0
-    }
-
-    /// Closes the descriptor an adoption refused, and records the rejection
-    /// only because the close ran.
-    ///
-    /// Verified rather than assumed: the GENERATION still matches, so a
-    /// recycled number cannot be closed in the original's place; and the close
-    /// goes through the token's injectable closer, so a test can witness that
-    /// this site invoked it. An earlier version probed the descriptor with
-    /// fcntl AFTER closing — racy, because by then the number may belong to
-    /// someone else, and it failed under load on unchanged code.
-    ///
-    /// Both a success-shaped substitute for the call and deleting it outright
-    /// leave the injected closer uninvoked, so both are caught.
-    static func closeRejectedAdoption(
-        _ fd: Int32, generation: Int, owner: SSHTransport.AuditToken?
-    ) {
-        let current = lock.withLock { generations[fd] ?? 0 }
-        guard current == generation else { return }   // recycled: not ours to close
-        if (owner?.closer ?? realClose)(fd) == 0 {
-            owner?.recordRejection()
-        }
-    }
-}
-
-#if canImport(Glibc)
-private func Glibc_close(_ fd: Int32) -> Int32 { Glibc.close(fd) }
-private func Glibc_shutdownSocket(_ fd: Int32) -> Int32 { Glibc.shutdown(fd, Int32(SHUT_RDWR)) }
-#else
-private func Glibc_close(_ fd: Int32) -> Int32 { Darwin.close(fd) }
-private func Glibc_shutdownSocket(_ fd: Int32) -> Int32 { Darwin.shutdown(fd, Int32(SHUT_RDWR)) }
-#endif
-
-/// Process-wide libssh2 init.
-///
-/// libssh2 requires this before any session call. It was previously only ever
-/// invoked by tests while a comment claimed callers got it "for free" — a
-/// documented behaviour that did not exist, so production paths ran
-/// uninitialised. `openSession` now calls it, and the return code is preserved
-/// rather than discarded.
-public enum SSHRuntime {
-    private static let lock = NSLock()
-    private static var started = false
-    private static var result: Int32 = 0
-
-    /// Idempotent and thread-safe. Throws if libssh2 refused to initialise.
-    public static func ensureStarted() throws {
-        lock.lock(); defer { lock.unlock() }
-        if !started {
-            result = libssh2_init(0)
-            started = true
-        }
-        guard result == 0 else { throw SSHError.runtimeInitFailed(code: result) }
-    }
-
-    /// Retained for callers that want to initialise eagerly.
-    public static func start() { try? ensureStarted() }
-}
