+++ Tests/HerdrKitTests/SessionRecoveryTests.swift
@@ -1283,17 +1283,6 @@ final class PaneSnapshotAccessTests: XCTestCase {
         #endif
     }
 
-    /// libssh2's header search path, from pkg-config.
-    ///
-    /// SwiftPM hands the CSSH target these flags automatically — but this test
-    /// spawns symbolgraph-extract ITSELF, so it inherits nothing. On Debian that
-    /// was invisible because /usr/include is a default search path; on macOS,
-    /// Homebrew's prefix is not, so the shim's `#include <libssh2.h>` failed and
-    /// the whole module could not be built for extraction.
-    ///
-    /// That is a defect the pkg-config fix CREATED: moving from a hardcoded path
-    /// to a resolved one fixed the build and broke every hand-rolled tool
-    /// invocation that had been relying on the hardcoded assumption.
     /// Every directory under `.build` that holds a `module.modulemap`, so
     /// symbolgraph-extract can resolve HerdrKit's TRANSITIVE C modules.
     ///
@@ -1313,24 +1302,6 @@ final class PaneSnapshotAccessTests: XCTestCase {
             dirs.insert(url.deletingLastPathComponent().path)
         }
         return dirs.sorted()
-    }
-
-    static func libssh2ClangFlags() -> [String] {
-        let p = Process()
-        p.executableURL = URL(fileURLWithPath: "/usr/bin/env")
-        p.arguments = ["pkg-config", "--cflags-only-I", "libssh2"]
-        let out = Pipe()
-        p.standardOutput = out
-        p.standardError = Pipe()
-        guard (try? p.run()) != nil else { return [] }
-        p.waitUntilExit()
-        let text = String(
-            data: out.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) ?? ""
-        return text.split(separator: " ").compactMap { flag -> [String]? in
-            let f = flag.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard f.hasPrefix("-I"), f.count > 2 else { return nil }
-            return ["-Xcc", f]
-        }.flatMap { $0 }
     }
 
     func testPaneSnapshotExposesNoInitialiserOutsideTheModule() throws {
@@ -1392,10 +1363,8 @@ final class PaneSnapshotAccessTests: XCTestCase {
             // compiler probe before adding.
             "-include-spi-symbols",
             "-I", moduleDir.path,
-            "-I", root.appendingPathComponent("Sources/CSSH").path,
         ]
         args += transitiveIncludes
-        args += Self.libssh2ClangFlags()
         args += Self.sdkFlags()
         args += ["-output-dir", output.path, "-minimum-access-level", "package"]
         process.arguments = args
